requirements.txt
python-telegram-bot>=20.0
neo4j>=5.0.0
python-dotenv
